requirements.txt
# Interfaz Gráfica
ttkbootstrap==1.10.1

# Manipulación de PDF
pypdf==4.1.0
pdf2image==1.17.0

# Procesamiento de Imágenes y Escáner
Pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.4

# Conversión de Documentos
docx2pdf==0.1.8
